Refactored app to use a map for state.books rather than an array
Changed how I am keeping track of my list of books. Originally, I was keeping it as an array, which made it difficult to work on specific books. I moved it to a map so I can access books by ID directly. To that end, I also needed to update the first few functions that act upon that list of books. 

In addition, I added some comments to class methods. 

I also added the ability for books on the search screen to know what their shelf is, if I already have the book on one of my shelves.